move logic into hooks file
